Added more intricate Download System
Now all downloads are localized in one folder "Downloads", with a subfolder for each provider for sorting. SteamCMD mods are downloaded, then moved to the "Downloads\SteamCMD\APPID" folder. SteamWebAPI mods are put directly in Downloads\SteamWebAPI.
Also addressed which folder will be opened based on 3 criterias: 1. If 2 providers are in the queue, the "Open" button will open the "Downloads" folder 2. if a Mod in queue is highlighted, it will open the directory that mod is in 3. If no mods are highlighted, and all use the same provider, it will open that specific provider folder