Added 'Mod Name' option for mod Folder
This was a requested feature: This option allows automatic renaming of the downloaded mod, so it uses the mod name instead of the ID. This way its easier to identify what mod it is. It will only use UTF-8 characters and remove characters that are illegal in file/folder names.