Rebuilt 'Settings' menu from the ground up
Rebuilt the settings menu from the ground up with QStackedWidget. It now has a sidebar with categories and all options are sorted. This will make it easier to add options and not have an overly cluttered settings menu. I added a fairly versatile separator helper function that can be used anywhere I want.